client: move VM and block execution logic to FullSync vm: add maxBlocks option to runBlockchain
