+++ packages/client/lib/sync/fetcher/blockfetcher.ts
@@ -1,45 +1,19 @@
-const level = require('level')
 import { Fetcher, FetcherOptions } from './fetcher'
 import { Block, BlockBodyBuffer } from '@ethereumjs/block'
 import { Peer } from '../../net/peer'
 import { EthProtocolMethods } from '../../net/protocol'
-import VM from '@ethereumjs/vm'
-import { DefaultStateManager } from '@ethereumjs/vm/dist/state'
-import { SecureTrie as Trie } from '@ethereumjs/trie'
 
 /**
  * Implements an eth/62 based block fetcher
  * @memberof module:sync/fetcher
  */
 export class BlockFetcher extends Fetcher {
-  public vm: VM
-
   /**
    * Create new block fetcher
    * @param {FetcherOptions}
    */
   constructor(options: FetcherOptions) {
     super(options)
-
-    if (!this.config.vm) {
-      const db = level('./statedir')
-      const trie = new Trie(db)
-
-      const stateManager = new DefaultStateManager({
-        common: this.config.common,
-        trie,
-      })
-
-      this.vm = new VM({
-        common: this.config.common,
-        blockchain: this.chain.blockchain,
-        stateManager,
-      })
-    } else {
-      this.vm = this.config.vm
-      //@ts-ignore blockchain has readonly property
-      this.vm.blockchain = this.chain.blockchain
-    }
   }
 
   /**
@@ -85,13 +59,7 @@ export class BlockFetcher extends Fetcher {
     blocks = blocks.map((b: Block) =>
       Block.fromValuesArray(b.raw(), { common: this.config.common })
     )
-    await this.chain.blockchain.initPromise
-    for (let i = 0; i < blocks.length; i++) {
-      const block: Block = blocks[i]
-
-      await this.chain.blockchain.putBlock(block)
-      await this.vm.runBlockchain()
-    }
+    await this.chain.blockchain.putBlocks(blocks)
     await this.chain.update()
   }
