+++ packages/vm/lib/index.ts
@@ -272,9 +272,9 @@ export default class VM extends AsyncEventEmitter {
    *
    * @param blockchain -  An [@ethereumjs/blockchain](https://github.com/ethereumjs/ethereumjs-vm/tree/master/packages/blockchain) object to process
    */
-  async runBlockchain(blockchain?: Blockchain): Promise<void> {
+  async runBlockchain(blockchain?: Blockchain, maxBlocks?: number): Promise<void> {
     await this.init()
-    return runBlockchain.bind(this)(blockchain)
+    return runBlockchain.bind(this)(blockchain, maxBlocks)
   }
 
   /**
